[exercise-3] refactor: refactors the method that calculates the account balance in AccountService by using a native query in the transactions repository instead of doing the calculation programmatically (#3)
The purpose of this change is to refactor the function which previously was called Calcbalance2 in AccountService. This function, given an account, was calculating its balance based on its transactions. In this commit, the method is renamed to calculateAccountBalance and is refactored by using a native query added in the Transaction repository that executes the calculation in the DB, improving performance and making the code more clear. Calculation time went down from average 100+ ms to average <5ms.